Extract SSH keys into files
Managing ssh keys in YAML files is rather unpleasant. This commit moves
the actual keys into individual files, and opens up the possibility that
one user might have more than one key (as in, for example, the okfn
user).

This currently involves a rather nasty hack, that means we don't
actually support multiple authorized_keys. Work is in the pipeline to
change this.