Listen on IPv6 by default
By default, it's perfectly safe to bind to the IPv6 interface by default
(avoiding all the stupid "ipv6only=yes"-but-only-once stuff you usually
have to do with nginx).

Linux will listen on IPv4 as well on this interfaces, so long as
net.ipv6.bindv6only is off, which it is by default.