Fix check_mk
The whole check_mk setup was completely broken, because agents were
running 1.2.x and the server was running 1.1.x. A call to

    check_mk -u -II

on the server, which is supposed to refresh the inventory, was just hard
crashing with:

    Traceback (most recent call last):
      File "/usr/share/check_mk/modules/check_mk.py", line 4189, in <module>
        checknames = inventorable_checktypes("all")
      File "/usr/share/check_mk/modules/check_mk.py", line 1753, in inventorable_checktypes
        print check_info[k][3]
    KeyError: 3

This commit installs check-mk 1.2 from a PPA on both server and clients,
although unfortunately some clients will be broken (those that don't run
Ubuntu Precise).

This is still better than it was, though.