postfix: Drastic cleanup of postfix role
The postfix role was a mess. This commit cleans up most of the rubbish
therein, and makes it possible to manage a wide range of postfix
configurations using only ansible variables, rather than replicating
each and every postfix config tree in this repository.

Postfix main.cf is now managed as a template in the postfix role,
allowing us to easily roll out security modifications to the role across
the infrastructure.